s02e01/requirements.txt
openai>=1.13.3
python-dotenv
requests
loguru